refactor: replace seek setInterval in Player with requestAnimationFrame hook
This is done because the setInterval implementation caused a slow but sure
memory leak in the Renderer over the course of multiple days.

The animationFrame implementation increases resource consumption
somewhat, as it is constantly drawing the seek bar.

A possible optimization could possibly be to throttle the draw
operations to the previous refresh rate of 500 ms.

* Prettier goes crazy in this as well, oh well.